Pass PartitionEntry directly to from_aligned_stack() in 7 example files
Remove intermediate `as *const () as u32` and `as usize as u32` casts
from ExternalPartitionMemory::from_aligned_stack() calls, leveraging
the IntoEntryAddr impl for PartitionEntry.

+++ kernel/examples/mpu_cached_dynamic_test.rs
@@ -136,7 +136,7 @@ fn main() -> ! {
             let base = stk.as_u32_slice().as_ptr() as u32;
             ExternalPartitionMemory::from_aligned_stack(
                 stk,
-                entry_fns[i] as usize as u32,
+                entry_fns[i],
                 MpuRegion::new(base, REGION_SZ, 0),
                 i as u8,
             )
